refactor: reorganize readability's logic into eligibility/lifecycle/loading-screen/utils
Splits the flat readability/*.ts files into folders by responsibility
(eligibility checks, apply/remove lifecycle, loading screen, dock message
bridge), renames files after their exported function, and renames the
apply/remove public API to applyReadability/removeReadability.
