+++ src/editor/store/actions.ts
@@ -17,10 +17,7 @@ import {
   removeEmptyRules,
 } from '@stylebot/css';
 
-import {
-  apply as applyReadability,
-  remove as removeReadability,
-} from '@stylebot/readability';
+import { applyReadability, removeReadability } from '@stylebot/readability';
 
 import {
   Style,
