+++ src/inject-css/index.ts
@@ -7,7 +7,7 @@ import { extractImports, pruneImportCache } from '@stylebot/css';
 // Bypasses the @stylebot/readability barrel, whose side-effectful SCSS
 // import defeats tree-shaking and would pull the whole apply/reader/Defuddle
 // stack into this document_start bundle just for this one heuristic.
-import { isReaderable } from '../readability/heuristics';
+import { isReaderable } from '../readability/eligibility';
 import { getStylesForPage } from '@stylebot/styles';
 import { StyleMap, TabMessage } from '@stylebot/types';
