--- src/readability/apply.ts
@@ -1,164 +1,0 @@
-import { initReader } from './reader';
-import {
-  shouldRunOnUrl,
-  isMediaWikiMainPage,
-  shouldWaitForFullLoad,
-} from './heuristics';
-import { showLoader, hideLoader } from './loader';
-import { cacheUrl, didUrlChange, revertToCachedDocument } from './cache';
-
-import { ReadabilityActiveChanged } from '@stylebot/types';
-
-// Tells the background to refresh the badge — carries no state itself,
-// it re-queries the live DOM instead of trusting a passed value.
-const reportChanged = (): void => {
-  const message: ReadabilityActiveChanged = { name: 'ReadabilityActiveChanged' };
-  chrome.runtime.sendMessage(message);
-};
-
-// Client-rendered pages can still be empty right after load — retry a few
-// times before giving up, so hydration has a chance to finish.
-export const RETRY_DELAYS_MS = [300, 600, 1200];
-
-// Bumped on every apply()/remove() so a retry left over from a superseded
-// attempt (e.g. the user toggled off while one was pending) is a no-op.
-let generation = 0;
-
-let pendingRetry: ReturnType<typeof setTimeout> | null = null;
-
-const clearPendingRetry = (): void => {
-  if (pendingRetry !== null) {
-    clearTimeout(pendingRetry);
-    pendingRetry = null;
-  }
-};
-
-// Checked once, not retried like initReader()'s heuristic — more waiting
-// won't change whether this is the wiki's main page.
-const startIfEligible = (myGeneration: number): void => {
-  if (isMediaWikiMainPage()) {
-    remove();
-    return;
-  }
-
-  run(myGeneration);
-};
-
-const run = async (myGeneration: number, attempt = 0): Promise<void> => {
-  if (myGeneration !== generation) {
-    return;
-  }
-
-  try {
-    await initReader();
-    reportChanged();
-  } catch (e) {
-    if (myGeneration !== generation) {
-      return;
-    }
-
-    if (attempt < RETRY_DELAYS_MS.length) {
-      pendingRetry = setTimeout(
-        () => run(myGeneration, attempt + 1),
-        RETRY_DELAYS_MS[attempt]
-      );
-    } else {
-      remove();
-    }
-  }
-};
-
-export const apply = async (forceApply = false): Promise<void> => {
-  if (window !== window.top) {
-    return;
-  }
-
-  // Prevent duplicate calls for the same url if not force applying
-  if (!forceApply && !didUrlChange()) {
-    return;
-  }
-
-  cacheUrl();
-
-  if (!shouldRunOnUrl()) {
-    remove();
-    return;
-  }
-
-  generation++;
-  const myGeneration = generation;
-
-  showLoader();
-
-  scheduleStart(myGeneration);
-};
-
-// Ceiling on the deferred-parse wait: if a stray resource never resolves,
-// parse anyway rather than leaving the themed loader up indefinitely.
-const FULL_LOAD_TIMEOUT_MS = 5000;
-
-const scheduleStart = (myGeneration: number): void => {
-  // Lazy-image sites: wait for window `load` so real srcs have resolved.
-  if (shouldWaitForFullLoad()) {
-    if (document.readyState === 'complete') {
-      startIfEligible(myGeneration);
-      return;
-    }
-
-    let started = false;
-    const start = (): void => {
-      if (started) {
-        return;
-      }
-      started = true;
-      startIfEligible(myGeneration);
-    };
-
-    const timeout = setTimeout(start, FULL_LOAD_TIMEOUT_MS);
-    window.addEventListener(
-      'load',
-      () => {
-        clearTimeout(timeout);
-        start();
-      },
-      { once: true }
-    );
-    return;
-  }
-
-  // DOMContentLoaded fires once; attach only while still loading.
-  if (document.readyState === 'loading') {
-    document.addEventListener('DOMContentLoaded', () => {
-      startIfEligible(myGeneration);
-    });
-  } else {
-    startIfEligible(myGeneration);
-  }
-};
-
-// Kept in sync with the `.stylebot-reader.closing` transition duration.
-const CLOSE_TRANSITION_MS = 250;
-
-export const remove = (): void => {
-  generation++;
-  clearPendingRetry();
-  hideLoader();
-  revertToCachedDocument();
-
-  const host = document.getElementById('stylebot-reader');
-  const panel = host?.shadowRoot?.querySelector<HTMLElement>('.stylebot-reader');
-
-  if (!panel) {
-    host?.remove();
-    reportChanged();
-    return;
-  }
-
-  // The original page is already back underneath — fade the panel out
-  // before detaching it instead of cutting away abruptly.
-  panel.classList.add('closing');
-  setTimeout(() => {
-    host?.remove();
-    reportChanged();
-  }, CLOSE_TRANSITION_MS);
-};
