+++ src/readability/index.ts
@@ -1,5 +1,4 @@
 import './scss/index.scss';
 
-export { apply, remove } from './apply';
-export { isReaderable } from './heuristics';
-export { SIZES, WIDTHS, nearestStepIndex } from './steps';
+export { applyReadability, removeReadability } from './lifecycle';
+export { isReaderable } from './eligibility';
