+++ src/readability/lifecycle/document-cache.ts
@@ -5,14 +5,23 @@ declare global {
   }
 }
 
+/**
+ * Records the current URL so a later call can tell whether it's stale.
+ */
 export const cacheUrl = (): void => {
   window.stylebotReaderUrl = window.location.href;
 };
 
+/**
+ * Whether the page has navigated since the last cacheUrl() call.
+ */
 export const didUrlChange = (): boolean => {
   return window.stylebotReaderUrl !== window.location.href;
 };
 
+/**
+ * Detaches the original page body and remembers its nodes for later restore.
+ */
 export const cacheDocument = (): void => {
   const nodes = Array.prototype.slice
     .call(document.body.childNodes)
@@ -22,6 +31,9 @@ export const cacheDocument = (): void => {
   nodes.forEach(node => node.remove());
 };
 
+/**
+ * Re-attaches the original page body nodes cached by cacheDocument().
+ */
 export const revertToCachedDocument = (): void => {
   if (window.stylebotReaderOriginalDocumentBodyElements) {
     window.stylebotReaderOriginalDocumentBodyElements.forEach(node => {
