+++ src/readability/lifecycle/index.ts
@@ -1,0 +1,2 @@
+export { applyReadability } from './apply-readability';
+export { removeReadability } from './remove-readability';
